requirements.txt
pandas==2.1.4
numpy==1.26.2
datasets==2.16.1
transformers==4.36.2
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
openpyxl==3.1.2
jsonlines==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0

